+++ Backend/app/models/order_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Numeric, ForeignKey
+from sqlalchemy import Column, Numeric, ForeignKey, Integer
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 import uuid
@@ -12,7 +12,7 @@ class OrderItem(Base, AuditMixin):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     order_id = Column(UUID(as_uuid=True), ForeignKey("orders.id"), nullable=False)
     material_id = Column(UUID(as_uuid=True), ForeignKey("materials.id"), nullable=False)
-    quantity_m3 = Column(Numeric(10, 2), nullable=False)
+    quantity_m3 = Column(Integer, nullable=False)
     unit_price = Column(Numeric(10, 2), nullable=False)
     subtotal = Column(Numeric(12, 2), nullable=False)
